// components/market/types.ts
export type TopItem = {
  id: string
  name: string
  symbol: string
  price: string
  change: string
  iconUrl: string
}

export type HotItem = {
  id: string
  symbol: string
  date: string
  leverage: string
  type: string
  volume: string
  iconUrl: string
}

export type OptionsItem = {
  id: string
  symbol: string
  date: string
  leverage: string
  type: string
  volume: string
  iconUrl: string
}